[ddm-admin-client] Remove sneaky tokio::spawn()s (#7507)
Changes to our `ddm-admin-client` wrapper library:

* `advertise_prefix` is now `advertise_prefixes`, and is a straight
pass-through to the underlying progenitor client (i.e., it no longer
`tokio::spawn()`s an infinite retry loop). I inlined the retry loop at
all the call sites. This is messy in sled-agent, and I stuck comments on
them about cleaning up. It also makes it super obvious that one of them
is not correct (as we noted in #7377).
* `enable_stats` got the same treatment (no more `tokio::spawn`; inline
the retry loop at its one callsite).
* Removed `advertise_tunnel_endpoint()` (it had no callers).

This gets rid of all the hidden `tokio::spawn()`s, which fixes #7378.