Add a regression test for #7674.
This adds a test where we attempt to insert a large number of samples
into a database that we should never be able to reach. In the current
structure of the code, this test should take a _really_ long time to
run, because for each sample we insert, we attempt to claim a handle to
the database. As the resolution will never succeed, we should pay the
maximum connection cost on each sample, and the test times out.