Activate background task at least once
Pulled out of #7544, this commit includes changes necessary for
`activate_background_task` to activate and wait for a background task to
complete at least once _after_ the function is called. Otherwise, a race
condition exists where:

1. caller A calls `activate_background_task`
2. that background task queries some state, and operates on that
3. caller B changes that state after the background task queried it, and
   also calls `activate_background_task`
4. the background task is still running from caller A's invocation, and
   caller B's call of `activate_background_task` waits for that to
   complete

In this scenario caller B won't have waited for the background task to
pick up the changes it made.