[meta] update Rust style edition to 2024 (#7591)
Lots of improvements in this edition -- see https://doc.rust-lang.org/nightly/edition-guide/rust-2024/rustfmt-formatting-fixes.html.

This is a pretty major change that will almost certainly cause
conflicts. But it also seems to be one that's necessary, and hard to
stage. For folks at Oxide, see
[instructions](https://rfd.shared.oxide.computer/rfd/500#merging_across_editions)
on how to make this easier.

I had to run `cargo fmt` twice to reach a fixpoint -- that's quite
strange.