Oximeter: Change `try_force_collect` to be fire-and-forget (#7694)
The main point of the PR is the first commit
(dbdae3e53cc8648a5f9af2bcba37c053132f0477): This changes Oximeter's
`try_force_collect` method to be a fire-and-forget request to take a
collection instead of blocking until a collection has completed. This
allows us to drop the internal response `token` that's very problematic
for the transition to multinode Clickhouse.

The rest of the PR is trying to deal with the fallout of this: we had
several tests that used `try_force_collect` under the (correct!)
assumption that it only returned after the collection was done. After
changing the behavior, I injected some delays inside Oximeter to try to
force test failures, and all of the tests using `try_force_collect`
_except_ `test_mgs_metrics` failed. `test_mgs_metrics` went out of its
way to handle flakes and races (h/t @hawkw!), so the bulk of this PR is
really "can we generalize what `test_mgs_metrics` does and then use it
in the other metrics tests". This led to the helper type I called
`MetricsQuerier`, and reworking other tests to have similar "retry this
query until I see the result I expect, giving up after some unreasonably
long time (60 seconds)".

I'm vaguely worried I've missed something and this will introduce new
test flakes. I'm not sure how to get more confidence on that front,
other than letting this run through CI (and being on the lookout for new
metrics flakes for a little while after it lands).