[reconfigurator] Executor internal cleanup around iterators (#7722)
Builds on #7713, and is followup from
https://github.com/oxidecomputer/omicron/pull/7713#discussion_r1978146721.
In #7652 I changed all the executor substeps to take iterators instead
of `&BTreeMap` references that no longer existed, but that introduced a
weird split where the top-level caller had to filter the blueprint down
to just the items that the inner functions expected. @smklein pointed
out one place where the inner code was being extra defensive, which was
just more confusing than anything.

This PR removes that split: the top-level executor now always passes a
full `&Blueprint` down, and the inner modules are responsible for doing
their own filtering as appropriate. To easy testing, I kept the versions
that take an iterator of already-filtered items as private `*_impl`
functions that the new functions-that-take-a-full-`Blueprint` themselves
call too.