Make generic versions of two functions (#7737)
Pull the check to see if a service is gone into a generic Nexus function
`is_internal_service_gone`, and use that to determine if a Pantry is
gone.

Separate the idea of attaching a disk to the pantry from attaching a
volume with a new `call_pantry_attach_for_volume` function. This was
required for the snapshot export work and was pulled out of that branch.