Config property for target system software release (#7518)
Fixes #7280. Provides external API endpoints and corresponding datastore
methods to get/set the current `target_release` config property.
Releases are designated by their semantic version, and must be uploaded
to the TUF repo depot prior to being set as the target release.

This PR does _not_ attempt to actually upgrade the rack to the current
target release. The reconfigurator changes to plan & execute such an
upgrade will be handled as a follow-up.