Verify UUIDs before performing switch port settings queries (#7742)
In #6462, it was noted that if you provided a bad uuid for a BGP config
when setting a BGP peer, the api would accept it without reporting and
error, which would consequently break subsequent interactions that
relied on BGP configuration data.

This was due to us not validating the uuid before applying subsequent
updates to the switch port configuration tables. This PR adds UUID
validation to all transactions in the switch port configuration database
queries.