[sled-agent] Merge disk/dataset/zone `PUT` endpoints into one (#7762)
This is a step on the way to fixing #7309 and #5086. It's largely a
mechanical change: Instead of three separate endpoints that respectively
receive an `OmicronPhysicalDisksConfig`, `DatasetsConfig`, and
`OmicronZonesConfig`, sled-agent now exposes a single `PUT
/omicron-config` endpoint that takes all three of these at once:

```rust
pub struct OmicronSledConfig {
    pub disks_config: OmicronPhysicalDisksConfig,
    pub datasets_config: DatasetsConfig,
    pub zones_config: OmicronZonesConfig,
}
```

The handler for this endpoint is just the old three endpoints in a
trenchcoat: it handles the disks, then datasets, then zones, stopping on
any step that fails.

The two subsequent followup PRs will be to:

1. Make `OmicronSledConfig` a single ledgered config with one overall
generation, instead of the three separate ledgers with three separate
generations it is on this PR
2. Make this endpoint asynchronous via an internal reconciler task
(requiring callers to poll other endpoints like `/inventory` to
determine its progress)

When running this on a4x2, the `PUT`s that happen during RSS were pretty
slow; here are logs from one of the sleds:

```
20:15:00.033Z INFO SledAgent (dropshot (SledAgent)): request completed
    file = /home/john/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/dropshot-0.16.0/src/server.rs:867
    latency_us = 38131453
    local_addr = [fd00:1122:3344:103::1]:12345
    method = PUT
    remote_addr = [fd00:1122:3344:101::1]:42044
    req_id = 224d5006-d73d-40d7-90c8-eac31c62211f
    response_code = 200
    uri = /omicron-config
20:15:27.779Z INFO SledAgent (dropshot (SledAgent)): request completed
    file = /home/john/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/dropshot-0.16.0/src/server.rs:867
    latency_us = 25794156
    local_addr = [fd00:1122:3344:103::1]:12345
    method = PUT
    remote_addr = [fd00:1122:3344:101::1]:41843
    req_id = 407cdf41-a45f-41d7-80c6-37337c61424f
    response_code = 200
    uri = /omicron-config
20:16:56.013Z INFO SledAgent (dropshot (SledAgent)): request completed
    file = /home/john/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/dropshot-0.16.0/src/server.rs:867
    latency_us = 57100538
    local_addr = [fd00:1122:3344:103::1]:12345
    method = PUT
    remote_addr = [fd00:1122:3344:101::1]:59875
    req_id = 420648bc-f8e2-47b4-9776-3093aa165cd7
    response_code = 200
    uri = /omicron-config

... snip a bunch of 503s waiting for time to sync ...

20:24:34.313Z INFO SledAgent (dropshot (SledAgent)): request completed
    file = /home/john/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/dropshot-0.16.0/src/server.rs:867
    latency_us = 386395044
    local_addr = [fd00:1122:3344:103::1]:12345
    method = PUT
    remote_addr = [fd00:1122:3344:101::1]:58404
    req_id = 9bde0c77-e41c-4624-92b8-c21f45cd7989
    response_code = 200
    uri = /omicron-config
```

That's a whopping 386 seconds (six and a half minutes) for the final
`PUT`. Far from ideal, but probably good enough to land knowing we're
going to do work to make it async in the near future?

Once the system was up and running, I enabled blueprint execution to
confirm that `PUT`s that don't do anything are much faster than that,
and they are:

```
21:53:26.821Z INFO SledAgent (dropshot (SledAgent)): request completed
    file = /home/john/.cargo/registry/src/index.crates.io-1949cf8c6b5b557f/dropshot-0.16.0/src/server.rs:867
    latency_us = 286378
    local_addr = [fd00:1122:3344:103::1]:12345
    method = PUT
    remote_addr = [fd00:1122:3344:103::5]:64092
    req_id = 40f45713-3e12-44d7-82bf-ae134b71e4f7
    response_code = 200
    uri = /omicron-config
```