db: define Abandoned state for sagas
Define an "abandoned" saga state. An abandoned saga will not begin to be
executed by any SEC. Technicians mark sagas as abandoned using omdb;
this requires the saga's current executor not to be running (otherwise
it could receive a state update from Steno that will clobber the
Abandoned state).

This commit defines the new state in the database schema and fixes up
the DB crates accordingly, but adds no affordances for applying the new
saga state or considering it when deciding what sagas to recover.