`BlueprintSledConfig`: One generation number for sled-agent instead of three (#7788)
This is the first half of #7774. I naively expected to do the blueprint
side and the sled-agent side together, but the blueprint side alone is
pretty big already, so here it is.

The primary changes are in `nexus-types`:

* Remove `BlueprintZonesConfig`, `BlueprintDatasetsConfig`, and
`BlueprintPhysicalDisksConfig`; each of these was a generation number +
an appropriate `IdMap`
* `OmicronSledConfig` now contains a single generation number and three
`IdMap`s

and everything else is fallout from that. Most of the changes are
mechanical (and I highly recommend viewing the diff ignoring whitespace;
it cuts a few hundred +/- lines out). The bits that are less mechanical
and deserve a closer view are:

* Database schema changes
* Blueprint diff changes (bikeshedding welcome on how to present the
single config generation number)

One decision I think is fine but is also worth double checking: as of
this PR, we still have to send 3 generation numbers to sled-agent in
`OmicronSledConfig`, so I'm replicating the single `BlueprintSledConfig`
generation for all three of the sub-configs. I think this is fine; in
practice if deployed, it means a sled might see its disks or datasets
generation number jump a few units up without any changes to the actual
disks or datasets, but I believe that's effectively a no-op other than
rewriting the ledger JSON.