(8/5) Reduce VMM reservation contention (#7533)
https://github.com/oxidecomputer/omicron/pull/7498 was introduced to
benchmark the cost of concurrent instance provisioning, and it
demonstrated that through contention, performance can be significantly
on the VMM reservation pathway.

This PR optimizes that pathway, by removing the VMM reservation
transaction, and instead replacing it with some non-transactional
queries:

1. First, we query to see if the VMM reservation has already succeeded
(for idempotency)
2. Next, we query for all viable sled targets and affinity information
(`sled_find_targets_query`)
3. After parsing that data and picking a sled, we call
`sled_insert_resource_query` to INSERT a desired VMM record, and to
re-validate our constraints.

This change significantly improves performance in the `vmm-reservation`
benchmark, while upholding the necessary constraints implicit to VMM
provisioning.