improved TUF artifact replication robustness (#7519)
Closes #7399.

Nexus now owns and maintains a generation number for the set of
artifacts the system wants to be fully replicated, which is used by Sled
Agent to prevent conflicts. The generation number is stored in a new
singleton table based on the existing db_metadata singleton. I wrote up
`docs/tuf-artifact-replication.adoc` to provide a top-level overview of
the system and some of the conflicts that this refactor seeks to
prevent.

The Sled Agent artifact store APIs are modified. Two new APIs exist for
getting and putting an "artifact configuration", which is the list of
wanted artifacts and its associated generation number. The list request
returns the current generation number as well, and the PUT and "copy
from depot" requests require an up-to-date generation number in the
query string. The delete API is removed in favor of Sled Agent managing
deletions on its own whenever the configuration is updated.