minor: refactor session create methods (#7827)
There should be no functional changes here, though the internal error
messages are now slightly different between saml login and local login,
where before they were the same. Ran into this while working #7339. This
logic (which I wrote originally and shuffled around in #7374) never
really made sense, and I figured it's good prep for #7818 and friends to
clean it up.

The core of the change here is updating two existing functions that
returned `Result<Option<User>, Error>` to just return `Result<User,
Error>` and move the logic about what we do when the user was `None`
inside each function. In both cases, when the user was `None` we ended
up with an `Error::Unauthenticated` anyway, so we can just do that a
moment earlier and eliminate a lot of misdirection.