Trust-Quorum: Messages, Config, Crypto
This is the initial commit of the "real" trust quorum code as described
in RFD 238. This commit provides some of the foundation needed to
implement the trust quorum reconfiguration protocol.

The protocol itself will be implemented in a `Node` type in a
[sans-io](https://sans-io.readthedocs.io/) style with property based
tests for the full protocol, similar to LRTQ. Async code will utilize
the protocol and forward messages over sprockets channels, and handle
requests from Nexus.

This initial code was split out to keep the PR small, although
it has been used in some preliminary protocol code already that will
come in a follow up PR.