omdb: Add exit-code flag to blueprints diff
The output of `omdb nexus blueprints diff` can be extremely long on
fully populated rack. When performing customer upgrades, the length of
this output makes it difficult to determine if there are any changes
present.

To simplify this check, add a `--exit-code` flag like the flag of the
same name for `git-diff(1)`. This will exit with 1 if changes are found,
or 0 if there are no changes.