[1/N] omdb db sagas: list running and inject fault (#7732)
Breaking apart #4378 and copying the structure of #7695, add `omdb db
saga` as a command and implement the following sub-commands:

    Usage: omdb db saga [OPTIONS] <COMMAND>

    Commands:
      running  List running sagas
        fault  Inject an error into a saga's currently running node

This addresses part of the minimum amount required during a release
deployment:

1. after quiescing (#6804), omdb can query if there are any running
sagas.

2. if those running sagas are stuck in a loop and cannot be drained
(#7623), and the release contains a change to the DAG that causes Nexus
to panic after an upgrade (#7730), then omdb can inject a fault into the
database that would cause that saga to unwind when the affected Nexus is
restarted

Note for 2, unwinding a saga that is stuck in this way may not be valid
if there were significant changes between releases.